services/discovery-service/pyproject.toml
[project]
name = "discovery-service"
version = "0.1.0"
description = "YouTube video discovery service with smart targeting"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.1",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.10.0",
    "google-api-python-client>=2.185.0",
    "google-auth>=2.36.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-pubsub>=2.31.1",
    "google-cloud-secret-manager>=2.21.0",
    "google-cloud-monitoring>=2.23.2",
    "httpx>=0.28.0",
    "python-json-logger>=3.2.0",
    "pyyaml>=6.0.0",
    "packaging>=24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestDeprecationWarning",
]
